+++ smart-home-simulation/src/main/java/model/Zone.java
@@ -4,14 +4,16 @@ import java.util.ArrayList;
 
 public class Zone {
 	private static ArrayList<Zone> zoneList = new ArrayList<Zone>();
-	private int currentZone; 
+	private int currentZone;
 	private String period;
 	private String initialPeriod;
 	private String finalPeriod;
-	private int desiredTemperature;
-	
-	public Zone() {}
-	public Zone(int currentZone, String period, String initialPeriod, String finalPeriod, int desiredTemperature) {
+	private double desiredTemperature;
+
+	public Zone() {
+	}
+
+	public Zone(int currentZone, String period, String initialPeriod, String finalPeriod, double desiredTemperature) {
 		this.currentZone = currentZone;
 		this.period = period;
 		this.initialPeriod = initialPeriod;
@@ -19,9 +21,13 @@ public class Zone {
 		this.desiredTemperature = desiredTemperature;
 		getZoneList().add(this);
 	}
+
+	/**
+	 * Print statement
+	 */
 	public void printZoneDetails() {
 		for (Zone zone : getZoneList()) {
-			System.out.println("CurrentZone is " + (zone.getCurrentZone()+ 1));
+			System.out.println("CurrentZone is " + (zone.getCurrentZone() + 1));
 			System.out.println("Period is " + zone.getPeriod());
 			System.out.println("Initial Time Period is " + zone.getInitialPeriod());
 			System.out.println("Final Time Period is " + zone.getFinalPeriod());
@@ -29,22 +35,46 @@ public class Zone {
 			System.out.println("-----------");
 		}
 	}
+
+	/**
+	 * Getter
+	 */
 	public static ArrayList<Zone> getZoneList() {
 		return zoneList;
 	}
+
+	/**
+	 * Getter
+	 */
 	public int getCurrentZone() {
 		return currentZone;
 	}
+
+	/**
+	 * Getter
+	 */
 	public String getPeriod() {
 		return period;
 	}
+
+	/**
+	 * Getter
+	 */
 	public String getInitialPeriod() {
 		return initialPeriod;
 	}
+
+	/**
+	 * Getter
+	 */
 	public String getFinalPeriod() {
 		return finalPeriod;
 	}
-	public int getDesiredTemperature() {
+
+	/**
+	 * Getter
+	 */
+	public double getDesiredTemperature() {
 		return desiredTemperature;
 	}
 }
